requirements.txt
biopython>=1.81
regex>=2023.0.0
